cleanup and include meta package into ApiSurfaceTest
