+++ sdks/java/extensions/sql/src/main/codegen/data/Parser.tdd
@@ -16,14 +16,14 @@
 
 {
   # Generated parser implementation class package and name
-  package: "org.apache.beam.sdk.extensions.sql.parser.impl",
+  package: "org.apache.beam.sdk.extensions.sql.impl.parser.impl",
   class: "BeamSqlParserImpl",
 
   # List of import statements.
   imports: [
     "org.apache.calcite.sql.validate.*",
     "org.apache.calcite.util.*",
-    "org.apache.beam.sdk.extensions.sql.parser.*",
+    "org.apache.beam.sdk.extensions.sql.impl.parser.*",
     "java.util.*"
   ]
