+++ sdks/java/extensions/sql/src/main/java/org/apache/beam/sdk/extensions/sql/BeamSqlCli.java
@@ -21,16 +21,16 @@ import java.util.ArrayList;
 import java.util.List;
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.annotations.Experimental;
+import org.apache.beam.sdk.extensions.sql.impl.parser.BeamSqlParser;
+import org.apache.beam.sdk.extensions.sql.impl.parser.ColumnConstraint;
+import org.apache.beam.sdk.extensions.sql.impl.parser.ColumnDefinition;
+import org.apache.beam.sdk.extensions.sql.impl.parser.SqlCreateTable;
 import org.apache.beam.sdk.extensions.sql.impl.planner.BeamQueryPlanner;
 import org.apache.beam.sdk.extensions.sql.impl.rel.BeamRelNode;
 import org.apache.beam.sdk.extensions.sql.impl.utils.CalciteUtils;
 import org.apache.beam.sdk.extensions.sql.meta.Column;
 import org.apache.beam.sdk.extensions.sql.meta.Table;
 import org.apache.beam.sdk.extensions.sql.meta.store.MetaStore;
-import org.apache.beam.sdk.extensions.sql.parser.BeamSqlParser;
-import org.apache.beam.sdk.extensions.sql.parser.ColumnConstraint;
-import org.apache.beam.sdk.extensions.sql.parser.ColumnDefinition;
-import org.apache.beam.sdk.extensions.sql.parser.SqlCreateTable;
 import org.apache.beam.sdk.extensions.sql.schema.BeamSqlRow;
 import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
