+++ sdks/java/extensions/sql/src/main/java/org/apache/beam/sdk/extensions/sql/impl/parser/ColumnDefinition.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.sdk.extensions.sql.parser;
+package org.apache.beam.sdk.extensions.sql.impl.parser;
 
 import java.util.Arrays;
 import org.apache.calcite.sql.SqlDataTypeSpec;
