+++ sdks/java/extensions/sql/src/test/java/org/apache/beam/sdk/extensions/sql/BeamSqlApiSurfaceTest.java
@@ -52,6 +52,8 @@ public class BeamSqlApiSurfaceTest {
         .includingClass(BeamSqlEnv.class)
         .includingPackage("org.apache.beam.sdk.extensions.sql.schema",
             getClass().getClassLoader())
+        .includingPackage("org.apache.beam.sdk.extensions.sql.meta",
+            getClass().getClassLoader())
         .pruningPrefix("java")
         .pruningPattern("org[.]apache[.]beam[.]sdk[.]extensions[.]sql[.].*Test")
         .pruningPattern("org[.]apache[.]beam[.]sdk[.]extensions[.]sql[.].*TestBase");
